Move waitForIt
